// src/types/global.d.ts
export interface TabBarIconProps {
  focused: boolean;
  color: string;
  size: number;
  routeName: string;
}

export interface Colors {
  background: string;
  card: string;
  primary: string;
  text: string;
  border: string;
  mutedText: string;
  danger: string;
  success: string;
  pureWhite: string;
  pureBlack: string;
  placeholder: string;
}